backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
email-validator>=2.3.0
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt>=4.0.0
python-multipart
python-dotenv
alembic
httpx
openai
groq
qrcode[pil]
Pillow>=10.0.0
starlette
mercadopago
reportlab
